made subproof subroutines work with compiler
